Add files via upload
I've been working on this project for a while and am just now uploading to Github.  File is able to take in a path to the images and a path to output images (that shows a blue bounding box around any detected barcodes for each image & has a txt file containing Serial Codes extracted from the barcodes).